+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+- [Breaking change] Rename FormItem to FormXField 
+- [Breaking change] Expose the library file only
+- Improve documentation
+- Fix example project setup
+- Add optional soft validation when updating and validating an item
 
 ## v1.0.0
 - Initial release